feat: create `FeeService` class (#3595)
* chore: add fee service class


* chore: fix tests


* chore: remove trailing comma


* chore: resolve conflicts


* chore: fix test


* chore: rename fee service


---------
