+++ packages/relay/src/lib/relay.ts
@@ -187,7 +187,7 @@ export class Relay {
       this.cacheService,
     );
 
-    this.debugImpl = new DebugImpl(this.mirrorNodeClient, logger, this.cacheService);
+    this.debugImpl = new DebugImpl(this.mirrorNodeClient, logger, this.cacheService, hapiService);
     this.adminImpl = new AdminImpl(this.cacheService);
 
     this.hbarSpendingPlanConfigService = new HbarSpendingPlanConfigService(
