+++ packages/relay/src/lib/services/feeService/IFeeService.ts
@@ -1,0 +1,15 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import { IFeeHistory, RequestDetails } from '../../types';
+import { JsonRpcError } from '../../errors/JsonRpcError';
+
+export interface IFeeService {
+  feeHistory: (
+    blockCount: number,
+    newestBlock: string,
+    rewardPercentiles: Array<number> | null,
+    requestDetails: RequestDetails
+  ) => Promise<IFeeHistory | JsonRpcError>;
+
+  maxPriorityFeePerGas: (requestDetails: RequestDetails) => Promise<string>;
+}
